# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MineVault Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /world/{name}:
    get:
      operationId: getWorld
      tags: [world]
      summary: Get or create a world
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: World data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/World"
...
  /store/buy:
    post:
      operationId: buyItem
      tags: [store]
      summary: Buy an item from the store
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyItemInput"
      responses:
        "200":
          description: Purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResult"
        "400":
          description: Purchase failed
...
  /monetization/request-task:
    post:
      operationId: requestMonetizationTask
      tags: [monetization]
      summary: Start an ad task (drill boost or cool down)
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonetizationTaskInput"
      responses:
        "200":
          description: Task started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonetizationTaskResult"
  /monetization/verify-task:
    post:
      operationId: verifyMonetizationTask
      tags: [monetization]
      summary: Verify ad task completion and claim reward
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTaskInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyTaskResult"
        "400":
          description: Task not ready or invalid
components:
  schemas:
...
    World:
      type: object
      required: [id, name, ownerId, blockData]
      properties:
        id:
          type: integer
        name:
          type: string
        ownerId:
          type: ["integer", "null"]
        blockData:
          type: array
          items:
            type: array
            items:
              type: string
...
    BuyItemInput:
      type: object
      required: [itemId, quantity]
      properties:
        itemId:
          type: string
        quantity:
          type: integer
    PurchaseResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        newGemBalance:
          type: ["integer", "null"]
...
    MonetizationTaskInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [drill_boost, cool_down]
    MonetizationTaskResult:
      type: object
      required: [success, token, expiresAt]
      properties:
        success:
          type: boolean
        token:
          type: string
        expiresAt:
          type: string
        adUrl:
          type: string
    VerifyTaskInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
    VerifyTaskResult:
      type: object
      required: [success, reward]
      properties:
        success:
          type: boolean
        reward:
          type: string
        message:
          type: string