# lib/api-zod/src/generated/types/buyItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MineVault Game API
 * OpenAPI spec version: 0.1.0
 */

export interface BuyItemInput {
  itemId: string;
  quantity: number;
}