# lib/api-zod/src/generated/types/world.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MineVault Game API
 * OpenAPI spec version: 0.1.0
 */

export interface World {
  id: number;
  name: string;
  /** @nullable */
  ownerId: number | null;
  blockData: string[][];
}